Refactor free function
